+++ changelog.md
@@ -7,6 +7,17 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [0.1.4] - 2018-12-22
+
+### Changed
+
+- Interface `JWT` has now no required types, all optional.
+
+- Function `FindSubjectFunction` now accepts an object with subject and
+  additional properties as return type.
+
+- Function `FunctionVerify` now accepts value `undefined` as return type.
+
 ## [0.1.3] - 2018-12-21
 
 ### Changed
@@ -48,7 +59,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Initial release
 
-[Unreleased]: https://github.com/revam/node-jwt-manager/compare/v0.1.3...HEAD
+[Unreleased]: https://github.com/revam/node-jwt-manager/compare/v0.1.4...HEAD
+[0.1.4]: https://github.com/revam/node-jwt-manager/compare/v0.1.3...v0.1.4
 [0.1.3]: https://github.com/revam/node-jwt-manager/compare/v0.1.2...v0.1.3
 [0.1.2]: https://github.com/revam/node-jwt-manager/compare/v0.1.1...v0.1.2
 [0.1.1]: https://github.com/revam/node-jwt-manager/compare/v0.1.0...v0.1.1